extracting constructor functions to a new module
