+++ lib/money/money.rb
@@ -2,6 +2,7 @@
 require "money/bank/variable_exchange"
 require "money/bank/single_currency"
 require "money/money/arithmetic"
+require "money/money/constructors"
 require "money/money/formatting"
 
 # "Money is any object or record that is generally accepted as payment for
@@ -15,6 +16,7 @@ require "money/money/formatting"
 # @see http://en.wikipedia.org/wiki/Money
 class Money
   include Money::Arithmetic, Money::Formatting, Comparable
+  extend Constructors
 
   # Raised when smallest denomination of a currency is not defined
   class UndefinedSmallestDenomination < StandardError; end
@@ -118,72 +120,6 @@ class Money
     # @attr_writer rounding_mode Use this to specify the rounding mode
     attr_writer :rounding_mode
 
-    # Create a new money object with value 0.
-    #
-    # @param [Currency, String, Symbol] currency The currency to use.
-    #
-    # @return [Money]
-    #
-    # @example
-    #   Money.empty #=> #<Money @fractional=0>
-    def empty(currency = default_currency)
-      @empty ||= {}
-      @empty[currency] ||= Money.new(0, currency).freeze
-    end
-    alias_method :zero, :empty
-
-
-    # Creates a new Money object of the given value, using the Canadian
-    # dollar currency.
-    #
-    # @param [Integer] cents The cents value.
-    #
-    # @return [Money]
-    #
-    # @example
-    #   n = Money.ca_dollar(100)
-    #   n.cents    #=> 100
-    #   n.currency #=> #<Money::Currency id: cad>
-    def ca_dollar(cents)
-      Money.new(cents, "CAD")
-    end
-
-    alias_method :cad, :ca_dollar
-
-
-    # Creates a new Money object of the given value, using the American dollar
-    # currency.
-    #
-    # @param [Integer] cents The cents value.
-    #
-    # @return [Money]
-    #
-    # @example
-    #   n = Money.us_dollar(100)
-    #   n.cents    #=> 100
-    #   n.currency #=> #<Money::Currency id: usd>
-    def us_dollar(cents)
-      Money.new(cents, "USD")
-    end
-
-    alias_method :usd, :us_dollar
-
-
-    # Creates a new Money object of the given value, using the Euro currency.
-    #
-    # @param [Integer] cents The cents value.
-    #
-    # @return [Money]
-    #
-    # @example
-    #   n = Money.euro(100)
-    #   n.cents    #=> 100
-    #   n.currency #=> #<Money::Currency id: eur>
-    def euro(cents)
-      Money.new(cents, "EUR")
-    end
-
-    alias_method :eur, :euro
   end
 
   def self.default_currency
