+++ spec/money/constructors_spec.rb
@@ -1,0 +1,65 @@
+# encoding: utf-8
+
+require "spec_helper"
+
+describe Money::Constructors do
+
+  describe "::empty" do
+    it "creates a new Money object of 0 cents" do
+      expect(Money.empty).to eq Money.new(0)
+    end
+
+    it "memoizes the result" do
+      expect(Money.empty.object_id).to eq Money.empty.object_id
+    end
+
+    it "memoizes a result for each currency" do
+      expect(Money.empty(:cad).object_id).to eq Money.empty(:cad).object_id
+    end
+
+    it "doesn't allow money to be modified for a currency" do
+      expect(Money.empty).to be_frozen
+    end
+  end
+
+
+  describe "::zero" do
+    subject { Money.zero }
+    it { is_expected.to eq Money.empty }
+  end
+
+
+  describe "::ca_dollar" do
+    it "creates a new Money object of the given value in CAD" do
+      expect(Money.ca_dollar(50)).to eq Money.new(50, "CAD")
+    end
+
+    it "is aliased to ::cad" do
+      expect(Money.cad(50)).to eq Money.ca_dollar(50)
+    end
+  end
+
+
+  describe "::us_dollar" do
+    it "creates a new Money object of the given value in USD" do
+      expect(Money.us_dollar(50)).to eq Money.new(50, "USD")
+    end
+
+    it "is aliased to ::usd" do
+      expect(Money.usd(50)).to eq Money.us_dollar(50)
+    end
+  end
+
+
+  describe "::euro" do
+    it "creates a new Money object of the given value in EUR" do
+      expect(Money.euro(50)).to eq Money.new(50, "EUR")
+    end
+
+    it "is aliased to ::eur" do
+      expect(Money.eur(50)).to eq Money.euro(50)
+    end
+  end
+
+
+end
